Create VRstartup
Here i was given a list of upcoming projects needing to be delivered by a fictional VR startup company. My main responsibility was to analyze the current talent pool to ensure each project can be completed efficiently and effectively. 

This was a good exercise in joining tables, as well as a chance to look at case statements used in a more complex setting