Create analyse_AI_systems_with_SQL
For this project, I acted as an analyst in the AI industry. Using a database based on the Epochai dataset, I investigated the rise of Generative AI systems, the major players in the space, and the systems themselves. 

The purpose of this study was to improve my prompting using AI. By conversing with Chat GPT 4.0, i was able to refine and generate effective queries that answered a few investigative questions concerning the rise of AI systems. It's also helped me to deepen my understanding of SQL logic.