Create state_climate_study
Using climate data from each state in the United States to practice using window functions. The data provided from the table state_climate shows the average annual temperature for each state – this is the average temperature of every day in all parts of the state for that year. See comments on code for notes on my findings